refactor (replay.py): Write Memory as an iterable dataclass rather than a NamedTuple.
